Torch and Pickle Safe Load Fixes (#8566)
Fixes #8355.

### Description

This modifies the use of `torch.load` to load weights-only everywhere.
This will change some behaviour in that data needs to be converted to
tensors where possible, ie. converting Numpy arrays to tensors.

This issue was reported by @h3rrr in
https://github.com/Project-MONAI/MONAI/security/advisories/GHSA-6vm5-6jv9-rjpj.

This will also replace uses of `pickle` for saving and loading with
`torch.save/load`. This also requires that Numpy arrays be converted but
otherwise uses the restriction of weights-only to enforce safe loading.

This issue was reported by @h3rrr in
https://github.com/Project-MONAI/MONAI/security/advisories/GHSA-p8cm-mm2v-gwjm.

### Types of changes
<!--- Put an `x` in all the boxes that apply, and remove the not
applicable items -->
- [x] Non-breaking change (fix or new feature that would not break
existing functionality).
- [ ] Breaking change (fix or new feature that would cause existing
functionality to change).
- [ ] New tests added to cover the changes.
- [ ] Integration tests passed locally by running `./runtests.sh -f -u
--net --coverage`.
- [ ] Quick tests passed locally by running `./runtests.sh --quick
--unittests --disttests`.
- [x] In-line docstrings updated.
- [ ] Documentation updated, tested `make html` command in the `docs/`
folder.

---------

Signed-off-by: Eric Kerfoot <17726042+ericspod@users.noreply.github.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: YunLiu <55491388+KumoLiu@users.noreply.github.com>