Diffusion Model Encoder has an output layer set in forward method and this leads to problems (#8578)
Fixes #8577 .

### Description

This pull request sets the out layer of DiffusionModelEncoder in the
init method. This requires the inclusion of input_shape parameter in the
init method to calculate the input dimension to the last linear layer.

The output spatial shape derivation is a bit baroque, but allows for the
otherwise not very pleasant odd spatial dimensions at input.


### Types of changes
- [x] Non breaking change (fix or new feature that would cause existing
functionality to change). **The tutorial that was failing will need to
provide this input parameter, but input_shape is now defaulted to the
notebook dims.**
- [x] Quick tests passed locally by running `./runtests.sh --quick
--unittests --disttests`.
- [x] In-line docstrings updated.

---------

Signed-off-by: Virginia Fernandez <virginia.fernandez@kcl.ac.uk>
Co-authored-by: Virginia Fernandez <virginia.fernandez@kcl.ac.uk>
Co-authored-by: YunLiu <55491388+KumoLiu@users.noreply.github.com>